+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 * Fixed an issue in the **openapi_codegen** command where it created duplicate functions name from the swagger file.
 * Fixed an issue in the **update-release-notes** command where the *update type* argument was not verified.
-* Fixed an issue in the **validate** command where a non-existing docker image presented and no error was showed.
-
+* Fixed an issue in the **validate** command where no error was raised in case a non-existing docker image was presented.
+* Fixed an issue in the **format** command where format failed when trying to update invalid Docker image.
 
 # v1.2.8
 * Added a new validation that checks in playbooks for the usage of `DeleteContext` in **validate** command.